backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
websockets>=12.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Redis
redis>=5.0.1

# Celery
celery>=5.3.6
flower>=2.0.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.5.2

# GPT-Researcher
gpt-researcher==0.3.5
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0

# Document Processing
pypdf>=4.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
beautifulsoup4>=4.12.0

# Vector Stores
faiss-cpu>=1.7.4
qdrant-client>=1.7.3

# HTTP Client
httpx>=0.27.0
requests>=2.31.0

# Utilities
aiofiles>=23.2.1
tenacity>=8.2.3

# Monitoring
prometheus-client>=0.19.0
